third_party/blink/web_tests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=workers-expected.txt

This is a testharness.js-based test.
[FAIL] Worker() in a shared worker
  assert_equals: expected "%C3%A5" but got "ReferenceError: Worker is not defined"
Harness: the test ran to completion.
